% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ages_as_ordered}
\alias{ages_as_ordered}
\alias{ages_as_ordered_factor}
\alias{ages_as_integer}
\title{Convert ages from/to ordered factor}
\usage{
ages_as_ordered_factor(data, age_col = "age")

ages_as_integer(data, age_col = "est.age")
}
\arguments{
\item{data}{A data.frame with a column corresopnding to \code{age_col} or
a vector of values}

\item{age_col}{Character. The name of the column that contains ages}
}
\value{
A data.frame with the values in \code{age_col} converted to an
ordered factor
}
\description{
In order for the machine learning models to properly predict ages, the
known ages should be converted to an ordered factor during model fitting.
This will ensure that the predict.* functions return age values that actually
make sense.
}
