% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_na_age}
\alias{assign_na_age}
\title{Simple function that returns NA values}
\usage{
assign_na_age(x)
}
\arguments{
\item{x}{Any vector of any length}
}
\value{
A vector the same length as x containing only NA values
}
\description{
A vector of NA will be returned that is the length of \code{x}
}
