% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_age_models.R
\name{calc_mse_}
\alias{calc_mse_}
\title{Backend helper function to compute MSE or RMSE}
\usage{
calc_mse_(data, age_col = "age", root = FALSE)
}
\arguments{
\item{data}{A data.frame as created by \code{\link{assign_ages}}}

\item{age_col}{Character. Name of the age column in \code{data}}

\item{root}{Logical. computer MSE (FALSE, default) or RMSE (TRUE)}
}
\description{
This function is the engine for \code{\link{calc_mse}} and
\code{\link{calc_rmse}}. It was only created to remove the \code{root}
argument from the user in the main \code{calc_mse} function
}
