% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_agelen_data}
\alias{check_agelen_data}
\alias{check_age_data}
\alias{check_length_data}
\title{Check for age/length data in the data being estimated or predicted}
\usage{
check_age_data(data, age_col)

check_length_data(data, size_col)
}
\arguments{
\item{data}{A data.frame}

\item{age_col}{Character. The column name for the age column in \code{data}}

\item{size_col}{Character. The column name for the size column in \code{data}}
}
\value{
NULL. An error will be called if age/length data is missing
}
\description{
These are just simple helper functions used within other functions that check
to make sure that ages and lengths are present in the data and stop the
fucntion call if they are missing
}
