% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_model_type}
\alias{check_model_type}
\title{Check the model type and return standardized version}
\usage{
check_model_type(model)
}
\arguments{
\item{model}{A character string naming the model}
}
\value{
A standardized version of the model name, or an error if
\code{model} doesn't exist yet
}
\description{
This is a non-exported function to check whether the model type specified is
available and return a standardized version of the model name. This
standardized version will then feed into a S3 method for the given model.
}
