% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{laa_data}
\alias{laa_data}
\alias{laa_data_low_n}
\alias{laa_data_low_age_n}
\alias{laa_data_few_ages}
\alias{spp_data}
\alias{spp_data_low_n}
\alias{spp_data_low_age_n}
\alias{spp_data_few_ages}
\title{Example length-at-age data}
\format{
## `laa_data`
A data.frame with 244 rows and 2 columns:
\describe{
  \item{spp}{Species, only applicable for spp_data_* data.frames}
  \item{age}{Age of individual}
  \item{length}{Length of individual (arbitrary units)}
}

## `laa_data_low_n`
A data.frame with 27 rows and 2 columns:

## `laa_data_low_age_n`
A data.frame with 74 rows and 2 columns:

## `laa_data_few_ages`
A data.frame with 49 rows and 2 columns:

## `spp_data`
A data.frame with 1022 rows and 3 columns:

## `spp_data_low_n`
A data.frame with 87 rows and 3 columns:

## `spp_data_low_age_n`
A data.frame with 160 rows and 3 columns:

## `spp_data_few_ages`
A data.frame with 261 rows and 3 columns:
}
\usage{
laa_data

laa_data_low_n

laa_data_low_age_n

laa_data_few_ages

spp_data

spp_data_low_n

spp_data_low_age_n

spp_data_few_ages
}
\description{
Simple age-structured population data with age and length records for each
individual. \code{laa_data} represents a well-sampled age-length dataset,
whereas \code{laa_data_low_n} is one with few total samples,
\code{laa_data_low_age_n} is one with few samples in some ages,
and \code{laa_data_few_ages} is a dataset with few age groups sampled.
Species specific datasets are similar, but with the prefix \code{laa_}
replaced by \code{spp_}. These datasets contain species specific
length-at-age data
}
\keyword{datasets}
