% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{length_data}
\alias{length_data}
\alias{spp_length_data}
\title{Example length data}
\format{
## length data
A data.frame with one column and 244 rows
\describe{
  \item{spp}{Species, only in spp_length_data}
  \item{length}{Length of individual (arbitrary units)}
}

## `spp_length_data`
A data.frame with 1022 rows and 2 columns:
}
\usage{
length_data

spp_length_data
}
\description{
Simple vector and data.frame containing length measurements. These are used
in examples for functions that assign ages.
}
\keyword{datasets}
