% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_laa_cols}
\alias{rename_laa_cols}
\title{Simple helper function to rename size and age column names to age and length}
\usage{
rename_laa_cols(
  data,
  size_col = "length",
  age_col = "age",
  num_col = NULL,
  goback = FALSE
)
}
\arguments{
\item{data}{Any data.frame with some columns representing age and size}

\item{size_col}{Character. The name of the column containing sizes}

\item{age_col}{Character. The name of the column containing ages}

\item{num_col}{Character. The name of the column containing number of
individuals}

\item{goback}{Logical. Reverse names once they've already been renamed}
}
\value{
A data.frame the same as \code{data}, but with names changed
}
\description{
Simple helper function to rename size and age column names to age and length
}
