% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spp_levels}
\alias{spp_levels}
\alias{is_spp_in_levels}
\alias{is_spp_in_data}
\alias{spp_level}
\alias{rm_spp_level}
\alias{add_spp_level}
\title{Check for species in columns and/or levels and add to levels if present}
\usage{
is_spp_in_levels(levels)

is_spp_in_data(data)

spp_level(levels)

rm_spp_level(levels)

add_spp_level(data, levels)
}
\arguments{
\item{levels}{The levels argument passed from \code{\link{make_halk}}}

\item{data}{A data.frame with length-at-age data}
}
\value{
A character vector of levels possibly with 'spp' or 'species' added
}
\description{
These helper functions just check to see if a species column exists in the
data (designated as 'spp' or 'species'). If one of those columns exists,
but the column name is not in the levels argument it will get added to
levels.
}
