% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wb_spp_data}
\alias{wb_spp_data}
\alias{wb_spp_laa_data}
\alias{wb_spp_length_data}
\title{Separate species, county, waterbody example length-at-age and length data}
\format{
## `wb_spp_laa_data`
A data.frame with 36,849 records and 5 columns
\describe{
  \item{spp}{Species}
  \item{county}{Arbitrary example county name}
  \item{waterbody}{Arbitrary example waterbody name nested within county}
  \item{age}{Age of individual, only in wb_spp_laa_data}
  \item{length}{Length of individual (arbitrary units)}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9182 rows and 4 columns.
}
\usage{
wb_spp_laa_data

wb_spp_length_data
}
\description{
Simple age-structured population with age and/or length records, but
expanded across multiple counties and waterbodies for tests and examples
in \code{\link{make_halk}} used with levels.
}
\keyword{datasets}
